refactor(shadervalidate): split pack validators into their own TUs
main.cpp was 1086 lines, over the 800-line ceiling. Move the three
per-authoring-model validators (validatePack / validateAnimationPack /
validateSurfacePack) into packvalidators.cpp and their shared compile/report
helpers into packvalidatorcommon.cpp, both behind small headers in the new
PlasmaZones::ShaderValidate namespace. main.cpp keeps only argument parsing,
dispatch, isPackDir, and the emit-preamble sidecar path (282 lines); the two
new units are 661 and 199 lines. No behavior change: all 26 overlay, 69
animation, and 21 surface packs still validate, and the path-confine security
guard moves verbatim.
